tools/llms-converter/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai==1.55.3
httpx==0.27.2
tiktoken>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0